Move tests.gof.test_opt's test types to tests.gof.utils
